\name{greeks-package}
\Rdversion{1.1}
\alias{greeks-package}
\docType{package}
\title{
Financial Options Tools
}
\description{
Collection of functions for dealing with option contracts.
}
\details{
\tabular{ll}{
Package: \tab greeks\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-05-29\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
Basic functions to calculate individual call and put pricing, as well as associated
\sQuote{greeks}.
}
\author{
Jeffrey A. Ryan

Maintainer: Jeffrey A. Ryan <jeff.a.ryan@gmail.com>
}
\references{
Greeks(finance) \url{http://en.wikipedia.org/wiki/Greeks_(finance)}
}
\keyword{ package }
