% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Malliavin_Geometric_Asian_Greeks.R
\name{Malliavin_Geometric_Asian_Greeks}
\alias{Malliavin_Geometric_Asian_Greeks}
\title{Computes the Greeks of an Asian option with the Malliavin Monte Carlo
Method in the Black Scholes model}
\usage{
Malliavin_Geometric_Asian_Greeks(
  initial_price = 100,
  exercise_price = 100,
  r = 0,
  time_to_maturity = 1,
  volatility = 0.3,
  dividend_yield = 0,
  payoff = "call",
  greek = c("fair_value", "delta", "rho", "vega", "theta", "gamma"),
  model = "black_scholes",
  lambda = 0.2,
  alpha = 0.3,
  jump_distribution = function(n) stats::rt(n, df = 3),
  steps = round(time_to_maturity * 252),
  paths = 10000,
  seed = 1,
  antithetic = FALSE
)
}
\arguments{
\item{initial_price}{\itemize{
\item initial price of the underlying asset, can also be a
vector
}}

\item{exercise_price}{\itemize{
\item strike price of the option, can also be a vector
}}

\item{r}{\itemize{
\item risk-free interest rate
}}

\item{time_to_maturity}{\itemize{
\item time to maturity in years
}}

\item{volatility}{\itemize{
\item volatility of the underlying asset
}}

\item{dividend_yield}{\itemize{
\item dividend yield
}}

\item{payoff}{\itemize{
\item the payoff function, either a string in ("call", "put",
"digital_call", "digital_put"), or a function
}}

\item{greek}{\itemize{
\item the Greek to be calculated
}}

\item{model}{\itemize{
\item the model to be chosen in ("black_scholes", "jump_diffusion")
}}

\item{lambda}{\itemize{
\item the lambda of the Poisson process in the jump-diffusion model
}}

\item{alpha}{\itemize{
\item the alpha in the jump-diffusion model influences the jump size
}}

\item{jump_distribution}{\itemize{
\item the distribution of the jumps, choose a function
which generates random numbers with the desired distribution
}}

\item{steps}{\itemize{
\item the number of integration steps
}}

\item{paths}{\itemize{
\item the number of simulated paths
}}

\item{seed}{\itemize{
\item the seed of the random number generator
}}

\item{antithetic}{\itemize{
\item if TRUE, antithetic random numbers will be chosen to
decrease variance
}}
}
\value{
Named vector containing the values of the Greeks specified in the
parameter \code{greek}.
}
\description{
Computes the Greeks of an Asian option with the Malliavin Monte Carlo
Method in the Black Scholes model
}
\examples{
Malliavin_Asian_Greeks(initial_price = 110, exercise_price = 100,
r = 0.02, time_to_maturity = 4.5, dividend_yield = 0.015, volatility = 0.22,
greek = c("fair_value", "delta", "rho"), payoff = "put")

}
